backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
fastapi==0.104.1
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
multidict==6.7.0
propcache==0.4.1
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
twilio==9.9.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
